Change TId for VisitTypeRepository

+++ Modules/VisitSchedule/DentalClinic.VisitSchedule.Infrastructure/Repositories/VisitTypeRepository.cs
@@ -5,7 +5,7 @@ using DentalClinic.VisitSchedule.Infrastructure.Context;
 
 namespace DentalClinic.VisitSchedule.Infrastructure.Repositories;
 
-internal class VisitTypeRepository : GenericRepositoryBase<VisitType, long>, IVisitTypeRepository
+internal class VisitTypeRepository : GenericRepositoryBase<VisitType, int>, IVisitTypeRepository
 {
     public VisitTypeRepository(VisitScheduleDbContext context) 
         : base(context)
